서버 코드 업데이트(Ver.3.1)
헤더 분기수신 부활
공 가속도 적용
체크 골 활성화
스코어 전송

+++ AirHockey_Server/term_project/term_project/Server.h
@@ -32,8 +32,8 @@ struct Point2D {
 
 //객체의 2차원 가속도
 struct Accel2D {
-	int accel_x;
-	int accel_y;
+	float accel_x;
+	float accel_y;
 };
 
 DWORD WINAPI updateClient(LPVOID arg);		//Server내부 업데이트 + 데이터 송신
